backend/requirements.txt
alembic==1.10.3
anyio==3.6.2
bcrypt==4.0.1
black==23.3.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
ecdsa==0.18.0
fastapi==0.95.1
filelock==3.11.0
greenlet==2.0.2
h11==0.14.0
identify==2.5.22
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pip==22.0.2
platformdirs==3.2.0
pre-commit==3.2.2
psycopg2==2.9.6
pyasn1==0.4.8
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
setuptools==59.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
virtualenv==20.21.0
